backend/requirements/base.txt
Django==3.2.15
djangorestframework==3.12.2
django-environ==0.8.1
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
django-cors-headers==3.7.0
Pillow==9.0.0
PyJWT==2.0.1
Faker==6.6.1
argon2-cffi==21.3.0
django-redis==4.12.1
stripe==2.56.0

djoser==2.1.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.4

celery==5.2.3
flower==1.0.0
django-celery-email==3.0.0

pytz==2021.3

